+++ bin/violin.rb
@@ -4,7 +4,8 @@ MYVERSION = "0.1"
 
 C2S = "c2s"
 BOOGIE = "Boogie"
-CLEANUP = true
+$cleanup = true
+$graph = false
 
 puts "Violin version #{MYVERSION}"
 
@@ -38,6 +39,8 @@ def prepare()
   rounds, rest = rest.partition{|a| a =~ /\/rounds:[0-9]+/}
   delays, rest = rest.partition{|a| a =~ /\/delayBound:[0-9]+/}
   m2s, rest = rest.partition{|a| a =~ /\/multitosingle/}
+  keep, rest = rest.partition{|a| a =~ /\/keepFiles/}
+  graph, rest = rest.partition{|a| a =~ /\/graph(Of)?Trace/}
 
   if sources.empty? then
   	puts "Please specify at least one Boogie source file (.bpl)."
@@ -66,7 +69,8 @@ def prepare()
   end
 
   m2s = !m2s.empty?
-
+  $cleanup = keep.empty?
+  $graph = !graph.empty?
   rest = rest * " "
 
   src = "#{File.basename(sources.last,'.bpl')}.comp.bpl"
@@ -145,7 +149,13 @@ def verify( src, args )
     puts output
     exit
   else
-    puts output
+    if $graph && output =~ /[1-9][0-9]* errors?/ then
+      File.open("#{src}.trace",'w'){|f| f.write(output) }
+      `boogie-trace-parser.rb #{src}.trace`
+      File.delete("#{src}.trace") if $cleanup
+    else
+      puts output
+    end
     puts "Finished in #{Time.now - t0}s."
   end 
   
@@ -153,9 +163,7 @@ def verify( src, args )
 end
 
 def cleanup( files )
-  if CLEANUP then
-    File.delete( *files )
-  end
+  File.delete( *files ) if $cleanup
 end
 
 src, rounds, delays, rest = prepare()
